packages/bot/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tgvault-bot"
version = "0.1.0"
description = "TGVault Telegram bot"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=20.0",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["bot*"]
